requirements.txt
aiomysql==0.0.20
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
fastapi==0.61.0
h11==0.9.0
httptools==0.1.1
idna==2.10
pycparser==2.20
pycryptodome==3.9.8
pydantic==1.6.1
PyJWT==1.7.1
PyMySQL==0.9.2
redis==3.5.3
requests==2.24.0
six==1.15.0
starlette==0.13.6
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
